# Fix RR results