This patch does the following:
1. Removes Latency result for stream test as it is not relevant
2. Fixes a bug where the ThroughputSummary instead of LatencySummary was being used for summarizing P99 latency in the result table

Before Patch

```
+----------------+------------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
|  RESULT TYPE   |  SCENARIO  | PARALLELISM | HOST NETWORK | SERVICE | MESSAGE SIZE | SAME NODE | DURATION | SAMPLES |     AVG VALUE      |
+----------------+------------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
| Stream Results | TCP_STREAM | 1           | false        | false   | 1024         | false     | 5        | 2       | 4847.530000 (Mb/s) |
+----------------+------------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
+-------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
| RESULT TYPE | SCENARIO | PARALLELISM | HOST NETWORK | SERVICE | MESSAGE SIZE | SAME NODE | DURATION | SAMPLES |     AVG VALUE      |
+-------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
| Rr Results  | TCP_RR   | 1           | false        | false   | 1024         | false     | 5        | 2       | 8934.885000 (OP/s) |
+-------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
+------------------------+------------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
|      RESULT TYPE       |  SCENARIO  | PARALLELISM | HOST NETWORK | SERVICE | MESSAGE SIZE | SAME NODE | DURATION | SAMPLES |   99%TILE VALUE    |
+------------------------+------------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
| Stream Latency Results | TCP_STREAM | 1           | false        | false   | 1024         | false     | 5        | 2       | 4847.530000 (usec) |
+------------------------+------------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
+--------------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
|    RESULT TYPE     | SCENARIO | PARALLELISM | HOST NETWORK | SERVICE | MESSAGE SIZE | SAME NODE | DURATION | SAMPLES |   99%TILE VALUE    |
+--------------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
| Rr Latency Results | TCP_RR   | 1           | false        | false   | 1024         | false     | 5        | 2       | 8934.885000 (usec) |
+--------------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+

```

After patch

```
+----------------+------------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
|  RESULT TYPE   |  SCENARIO  | PARALLELISM | HOST NETWORK | SERVICE | MESSAGE SIZE | SAME NODE | DURATION | SAMPLES |     AVG VALUE      |
+----------------+------------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
| Stream Results | TCP_STREAM | 1           | false        | false   | 1024         | false     | 5        | 2       | 4922.180000 (Mb/s) |
+----------------+------------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
+-------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
| RESULT TYPE | SCENARIO | PARALLELISM | HOST NETWORK | SERVICE | MESSAGE SIZE | SAME NODE | DURATION | SAMPLES |     AVG VALUE      |
+-------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
| Rr Results  | TCP_RR   | 1           | false        | false   | 1024         | false     | 5        | 2       | 8853.965000 (OP/s) |
+-------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+--------------------+
+--------------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+-------------------+
|    RESULT TYPE     | SCENARIO | PARALLELISM | HOST NETWORK | SERVICE | MESSAGE SIZE | SAME NODE | DURATION | SAMPLES | AVG 99%TILE VALUE |
+--------------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+-------------------+
| RR Latency Results | TCP_RR   | 1           | false        | false   | 1024         | false     | 5        | 2       | 176.500000 (usec) |
+--------------------+----------+-------------+--------------+---------+--------------+-----------+----------+---------+-------------------+

```